Rename modules named *constraint to *filter.
