+++ dbt_semantic_interfaces/dbt_semantic_interfaces/test/objects/constraints/test_filter_renderer.py
@@ -1,6 +1,6 @@
 import logging
 
-from dbt_semantic_interfaces.objects.constraints.filter_renderer import (
+from dbt_semantic_interfaces.objects.where_filter.filter_renderer import (
     FilterRenderer,
     FilterCallParameterSets,
     DimensionCallParameterSet,
