+++ metricflow/plan_conversion/time_spine.py
@@ -9,7 +9,7 @@ from typing import List, Tuple
 import pandas as pd
 
 from metricflow.protocols.sql_client import SqlClient
-from metricflow.constraints.time_constraint import TimeRangeConstraint
+from metricflow.filters.time_constraint import TimeRangeConstraint
 from metricflow.dataflow.sql_table import SqlTable
 from metricflow.time.time_constants import ISO8601_PYTHON_FORMAT
 from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
