+++ metricflow/test/query/test_query_parser.py
@@ -4,7 +4,7 @@ import textwrap
 import pytest
 
 from dbt_semantic_interfaces.parsing.objects import YamlConfigFile
-from metricflow.constraints.time_constraint import TimeRangeConstraint
+from metricflow.filters.time_constraint import TimeRangeConstraint
 from metricflow.errors.errors import UnableToSatisfyQueryError
 from metricflow.plan_conversion.time_spine import TimeSpineSource
 from metricflow.specs import (
